pyproject.toml
[project]
name = "Washing-Machines-Anomaly-Detection"
version = "0.1.0"
description = "Microservices architecture"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "pyarrow>=22.0.0",
]

[dependency-groups]

ingestion = [
    "quixstreams>=3.0.0",
    "confluent-kafka>=2.4,<2.12",
    "pandas>=2.3.3",
]


data-engineering = [
    "pyspark>=4.1.1",
    "pyyaml>=6.0.3",
]

training_ml = [
    "numpy>=2.4.1",
    "scikit-learn>=1.8.0",
    "pandas>=2.2.0",
    "joblib>=1.4.0",
    "mlflow>=2.10.0",
]
feature-store = [
    "feast[redis,duckdb]>=0.58.0",
]
feature_store_apply = [
    "feast[redis,duckdb]>=0.58.0",
]
dataset-creation-2 = [
    "scikit-learn>=1.8.0",
    "sdv>=1.33.0",
]
dataset-creation = [
    "pyspark>=4.1.1",
]
spark-services = [
    "pyspark>=4.1.1",
    "pyyaml>=6.0.3",
]
training= [
    "mlflow>=3.9.0",
    "numpy>=2.4.1",
    "pandas>=2.3.3",
    "pydantic-settings>=2.11.0",
    "scikit-learn>=1.8.0",
]
mlflow = [
    "mlflow>=3.9.0",
]
streaming = [
    "quixstreams>=3.0.0",
    "pandas>=2.2.0",
    "scikit-learn>=1.5.0",
    "joblib>=1.4.0",
    "feast[redis]>=0.58.0",
    "python-dotenv>=1.0.0",
]

mlflow-server = [
    "mlflow>=2.10.0",
    "boto3>=1.34.0",
    "psycopg2-binary>=2.9.0", 
]
inference = [
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "joblib>=1.3.0",
    "mlflow>=2.10.0",
    "feast[redis]>=0.30.0",
    "python-dotenv>=1.0.0",
    "confluent-kafka>=2.11.1",
    "quixstreams>=3.23.3",
    "numpy>=2.4.2",
]
rag-service = [
    "fastembed>=0.7.4",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=2.0,<3.0",
    "langchain-qdrant>=1.1.0",
    "langchain-text-splitters>=1.1.1",
    "pydantic-settings>=2.11.0",
]
mcp-server-service = [
    "confluent-kafka>=2.11.1",
    "fastapi>=0.129.0",
    "fastembed>=0.7.4",
    "flashrank>=0.2.10",
    "langchain-google-genai>=2.1.12",
    "langchain-qdrant>=1.1.0",
    "openai>=2.24.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.11.0",
    "uvicorn>=0.34.0",
]
offline-files = [
    "pandas>=2.3.3",
    "pyarrow>=22.0.0",
]
cold-start = [
    "feast[redis]>=0.60.0",
]
retraining = [
    "feast>=0.60.0",
    "mlflow>=3.9.0",
    "numpy>=2.4.2",
    "pandas>=2.3.3",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.11.0",
    "redis>=4.6.0",
    "scikit-learn>=1.8.0",
]
if-anomaly = [
    "pydantic-settings>=2.11.0",
    "quixstreams>=3.23.3",
    "requests>=2.32.5",
]
langchain-app = [
    "fastapi>=0.129.0",
    "httpx>=0.28.1",
    "langchain>=1.2.10",
    "langchain-core>=1.2.15",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-openai>=1.1.10",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "uvicorn>=0.34.0",
]
mcp-server = [
    "fastapi>=0.129.0",
    "fastembed>=0.7.4",
    "fastmcp>=2.13.0.2",
    "langchain-google-genai>=2.1.12",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "qdrant-client>=1.17.0",
    "sentence-transformers>=5.2.3",
    "uvicorn>=0.34.0",
]

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true
